Added new .jpg for back of card design.
Also, that last commit was entirely by accident. I dropped my phone on my laptop and it submitted it before I could give a description. The last commit was actually 52 .jpg's that I created in photoshop (As with the back of the card design) as to be used in the laker's rummy game :) Sorry about that inconvenience.